Setting seed properly

+++ src/encoder-recoders-decoder-rlnc.h
@@ -43,7 +43,7 @@ public:
     // Encoder creation and settings
     m_encoder = encoder_factory.build ();
     m_encoder->set_systematic_off ();
-    m_encoder->seed (time (0));
+    m_encoder->seed ((uint32_t)time (0));
 
     // Initialize the input data
     std::vector<uint8_t> data (m_encoder->block_size (), 'x');
